Don't limit hello intervals to 65 seconds.
The protocol allows them to go up to 650 seconds, the value in msec
cannot be a short.

+++ network.h
@@ -30,7 +30,7 @@ struct buffered_update {
 struct network_conf {
     char *ifname;
     unsigned short cost;
-    unsigned short hello_interval;
+    unsigned hello_interval;
     char wired;
     char split_horizon;
     char lq;
@@ -77,7 +77,7 @@ struct network {
     unsigned int bucket;
     time_t activity_time;
     unsigned short hello_seqno;
-    unsigned short hello_interval;
+    unsigned hello_interval;
 };
 
 #define NET_CONF(_net, _field) \
